rpc,server,testutils: move DRPC cluster setting to rpcbase
Epic: CRDB-48935
Informs: None
Release note: None
