+++ pkg/rpc/rpcbase/BUILD.bazel
@@ -25,7 +25,10 @@ go_library(
         "//pkg/keys",
         "//pkg/roachpb",
         "//pkg/rpc/rpcpb",
+        "//pkg/settings",
+        "//pkg/util/buildutil",
         "//pkg/util/envutil",
+        "@com_github_cockroachdb_errors//:errors",
         "@io_storj_drpc//:drpc",
         "@org_golang_google_grpc//:grpc",
     ],
