+++ pkg/server/start_listen.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/cockroachdb/cmux"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
+	"github.com/cockroachdb/cockroach/pkg/rpc/rpcbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/netutil"
@@ -137,8 +138,8 @@ func startListenRPCAndSQL(
 	// Host drpc only if it's _possible_ to turn it on (this requires a test build
 	// or env var). If the setting _is_ on, then it was overridden in testing and
 	// we want to host the server too.
-	hostDRPC := rpc.ExperimentalDRPCEnabled.Validate(nil /* not used */, true) == nil ||
-		rpc.ExperimentalDRPCEnabled.Get(&cfg.Settings.SV)
+	hostDRPC := rpcbase.ExperimentalDRPCEnabled.Validate(nil /* not used */, true) == nil ||
+		rpcbase.ExperimentalDRPCEnabled.Get(&cfg.Settings.SV)
 
 	// If we're not hosting drpc, make a listener that never accepts anything.
 	// We will start the dRPC server all the same; it barely consumes any
